Revert "add cmake build system, split code and better rev_comp"
